Rename module hotfix
