--- ModuleDriverCarService/pom.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com</groupId>
-        <artifactId>taxiApp</artifactId>
-        <version>0.0.1-SNAPSHOT</version>
-    </parent>
-
-    <groupId>com.kpavlov</groupId>
-    <artifactId>ModuleDriverCarService</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <name>ModuleDriverCarService</name>
-    <description>ModuleDriverCarService</description>
-    <url/>
-
-</project>
