--- ModuleDriverCarService/src/main/java/com/kpavlov/moduledrivercarservice/exceprion/ResourceNotFoundException.java
@@ -1,8 +1,0 @@
-package com.kpavlov.moduledrivercarservice.exceprion;
-
-public class ResourceNotFoundException extends RuntimeException{
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
