--- ModuleDriverCarService/src/main/java/com/kpavlov/moduledrivercarservice/model/Car.java
@@ -1,34 +1,0 @@
-package com.kpavlov.moduledrivercarservice.model;
-
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@Entity
-@Table(name = "car")
-public class Car {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String model;
-
-    @Column(nullable = false)
-    private String registrationCode;
-
-    @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private CarStatus status;
-}
