--- ModuleDriverCarService/src/main/java/com/kpavlov/moduledrivercarservice/util/ErrorMessages.java
@@ -1,18 +1,0 @@
-package com.kpavlov.moduledrivercarservice.util;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class ErrorMessages {
-
-    public static final String VALIDATION_FAILED_MESSAGE = "Validation failed";
-
-    public static final String HTTP_MESSAGE_NOT_READABLE_MESSAGE = "Request body is missing or could not be read";
-
-    public static final String INTERNAL_SERVER_ERROR_MESSAGE = "An unexpected error occurred.";
-
-    public static final String NOT_FOUND_MESSAGE = "%s %s not found.";
-
-    public static final String ALREADY_USED_MESSAGE = "Duplicate or already used value: %s";
-}
