--- ModuleDriverCarService/src/main/resources/application.yml
@@ -1,12 +1,0 @@
-server:
-  port: 8081
-
-spring:
-  datasource:
-    driver-class-name: org.postgresql.Driver
-    url: jdbc:postgresql://localhost:5433/taxi_app_db
-    username: postgres
-    password: postgrespass
-
-  liquibase:
-    change-log: classpath:db/changelog/main-changelog.xml
