+++ driver-service/src/main/java/com/kpavlov/moduledrivercarservice/controller/CarController.java
@@ -1,0 +1,70 @@
+package com.kpavlov.moduledrivercarservice.controller;
+
+import com.kpavlov.moduledrivercarservice.dto.response.CarResponse;
+import com.kpavlov.moduledrivercarservice.model.Car;
+import com.kpavlov.moduledrivercarservice.repository.CarRepository;
+import com.kpavlov.moduledrivercarservice.service.CarService;
+import com.kpavlov.moduledrivercarservice.dto.request.create.CarCreateRequest;
+import com.kpavlov.moduledrivercarservice.dto.request.update.CarUpdateRequest;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.PathVariable;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/api/v1/cars")
+public class CarController {
+
+    private final CarService carService;
+    private final CarRepository carRepository;
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public CarResponse createCar(@RequestBody @Valid CarCreateRequest createCarRequest) {
+        return carService.createCar(createCarRequest);
+    }
+
+    @PutMapping("/{id}")
+    public CarResponse updateCar(
+            @PathVariable Long id,
+            @RequestBody @Valid CarUpdateRequest updateCarRequest) {
+        return carService.updateCar(id, updateCarRequest);
+    }
+
+    @GetMapping("/{id}")
+    public CarResponse getById(@PathVariable Long id) {
+        return carService.getCarById(id);
+    }
+
+    @DeleteMapping("/{id}/hard")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteCar(@PathVariable Long id) {
+        carService.deleteCar(id);
+    }
+
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void softDeleteCar(@PathVariable Long id) {
+        carService.softDeleteCar(id);
+    }
+
+    @GetMapping
+    public Page<Car> getAllCars(@RequestParam(value = "offset", defaultValue = "0") @Min(0) Integer offset,
+                                @RequestParam(value = "limit", defaultValue = "20") @Min(1) @Max(100) Integer limit) {
+        return carRepository.findAll(PageRequest.of(offset, limit));
+    }
+}
