+++ driver-service/src/main/java/com/kpavlov/moduledrivercarservice/dto/response/error/MultiErrorResponse.java
