+++ driver-service/src/main/java/com/kpavlov/moduledrivercarservice/exceprion/CarNotFoundException.java
@@ -1,0 +1,8 @@
+package com.kpavlov.moduledrivercarservice.exceprion;
+
+public class CarNotFoundException extends RuntimeException {
+
+    public CarNotFoundException(String message) {
+        super(message);
+    }
+}
