+++ driver-service/src/main/java/com/kpavlov/moduledrivercarservice/exceprion/DriverNotFoundException.java
@@ -1,0 +1,8 @@
+package com.kpavlov.moduledrivercarservice.exceprion;
+
+public class DriverNotFoundException extends RuntimeException{
+
+    public DriverNotFoundException(String message) {
+        super(message);
+    }
+}
