+++ driver-service/src/main/java/com/kpavlov/moduledrivercarservice/model/Car.java
@@ -1,0 +1,63 @@
+package com.kpavlov.moduledrivercarservice.model;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@Entity
+@Table(name = "car")
+public class Car {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String model;
+
+    @Column(nullable = false)
+    private String registrationCode;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private CarStatus status;
+
+    public String getRegistrationCode() {
+        return registrationCode;
+    }
+
+    public void setRegistrationCode(String registrationCode) {
+        this.registrationCode = registrationCode;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getModel() {
+        return model;
+    }
+
+    public CarStatus getStatus() {
+        return status;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public void setModel(String model) {
+        this.model = model;
+    }
+
+    public void setStatus(CarStatus status) {
+        this.status = status;
+    }
+}
