+++ driver-service/src/main/java/com/kpavlov/moduledrivercarservice/model/CarStatus.java
