+++ driver-service/src/main/java/com/kpavlov/moduledrivercarservice/util/ErrorMessages.java
@@ -1,0 +1,14 @@
+package com.kpavlov.moduledrivercarservice.util;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class ErrorMessages {
+
+    public static final String VALIDATION_FAILED_MESSAGE = "Validation failed";
+
+    public static final String HTTP_MESSAGE_NOT_READABLE_MESSAGE = "Request body is missing or could not be read";
+
+    public static final String INTERNAL_SERVER_ERROR_MESSAGE = "An unexpected error occurred.";
+}
