+++ driver-service/src/main/resources/db/changelog/main-changelog.xml
@@ -1,0 +1,16 @@
+<databaseChangeLog xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                   xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+                   http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.0.xsd">
+
+    <changeSet id="1" author="strangerforfallen@gmail.com">
+        <tagDatabase tag="db_init"/>
+    </changeSet>
+
+    <include file="v1/v1-accumulate.xml" relativeToChangelogFile="true"/>
+
+    <changeSet id="2" author="strangerforfallen@gmail.com" >
+        <tagDatabase tag="v_1.0.0"/>
+    </changeSet>
+
+</databaseChangeLog>
