+++ driver-service/src/main/resources/messages.properties
@@ -1,0 +1,10 @@
+error.duplicate.phone=The phone number %s is already in use.
+error.duplicate.email=The phone number %s is already in use.
+error.duplicate.regCode=The code number %s is already in use.
+error.not.found=Item with ID %d not found.
+field.required=Field {0} must be filled with value
+field.null=Field {0} cannot be null
+field.size=Field size {0} must be between {2} ? {1}
+field.phone=Phone must be real, example: +375291234567
+field.email=Email must be real, example: asd@gmail.com
+field.code=Code must be real, example: 1234AA-1
